split out response related stuff to it's own header
