+++ src/Integrity/IntegritySession.h
@@ -19,13 +19,6 @@
 
 #pragma once
 
-#include <string>
-#include <vector>
-#include <set>
-#include <list>
-#include <memory>
-#include <future>
-
 extern "C" {
 #include <mksapi.h>
 }
@@ -98,122 +91,3 @@ public:
 private:
 	wchar_t *name, *app;
 };
-
-class IntegrityResponse {
-public:
-	IntegrityResponse(mksCmdRunner commandRunner, mksResponse response, const IntegrityCommand& command) 
-		: command(command)
-	{
-		this->commandRunner = commandRunner;
-		this->response = response;
-	}
-
-	~IntegrityResponse() 
-	{
-		// releaseing the command runner also releases the response
-		if (commandRunner != NULL) {
-			mksReleaseCmdRunner(commandRunner);
-			commandRunner = NULL;
-			response = NULL;
-		}
-	}
-
-	mksAPIException getException() const {
-		if (response != NULL) {
-			return mksResponseGetAPIException(response);
-		} else {
-			return NULL;
-		}
-	}
-
-	int getExitCode() {
-		int exitCode = -1;
-		if (response != NULL) {
-			mksResponseGetExitCode(response, &exitCode);
-		}
-		return exitCode;
-	}
-
-	mksResponse getResponse() const {
-		return response;
-	}
-
-	const IntegrityCommand& getCommand() const {
-		return command;
-	}
-
-private:
-	mksCmdRunner commandRunner;
-	mksResponse response;
-	IntegrityCommand command;
-};
-
-class WorkItemIterator {
-public:
-	WorkItemIterator(mksResponse response, mksWorkItem item)  : item(item), response(response){
-	}
-
-	WorkItemIterator& operator++() {
-		item = mksResponseGetNextWorkItem(response);
-		return *this;
-	}
-
-	bool operator!=(const WorkItemIterator& other) const {
-		return item != other.item;
-	}
-
-	mksWorkItem operator*() const {
-		return item;
-	}
-
-private:
-	mksWorkItem item;
-	mksResponse response;
-};
-
-inline WorkItemIterator begin(mksResponse response) {
-	return WorkItemIterator(response, mksResponseGetFirstWorkItem(response));
-}
-
-inline WorkItemIterator end(mksResponse response) {
-	return WorkItemIterator(response, NULL);
-}
-
-inline WorkItemIterator begin(const IntegrityResponse& response) {
-	return begin(response.getResponse());
-}
-
-inline WorkItemIterator end(const IntegrityResponse& response) {
-	return end(response.getResponse());
-}
-
-class WorkItemFieldIterator {
-public:
-	WorkItemFieldIterator(mksWorkItem item, mksField field) : field(field), item(item){
-	}
-
-	WorkItemFieldIterator& operator++() {
-		field = mksWorkItemGetNextField(item);
-		return *this;
-	}
-
-	bool operator!=(const WorkItemFieldIterator& other) const {
-		return field != other.field;
-	}
-
-	mksField operator*() const {
-		return field;
-	}
-
-private:
-	mksWorkItem item;
-	mksField field;
-};
-
-inline WorkItemFieldIterator begin(mksWorkItem item) {
-	return WorkItemFieldIterator(item, mksWorkItemGetFirstField(item));
-}
-
-inline WorkItemFieldIterator end(mksWorkItem item) {
-	return WorkItemFieldIterator(item, NULL);
-}
