+++ src/TortoiseShell/TortoiseShell.vcxproj.filters
@@ -102,6 +102,9 @@
     <ClCompile Include="..\Utils\CrashReport.cpp">
       <Filter>Utils</Filter>
     </ClCompile>
+    <ClCompile Include="..\Integrity\IntegrityResponse.cpp">
+      <Filter>Integrity</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="ShellExt.def">
@@ -201,6 +204,9 @@
     </ClInclude>
     <ClInclude Include="..\Utils\EventLog.h">
       <Filter>Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="..\Integrity\IntegrityResponse.h">
+      <Filter>Integrity</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
